went over how to export and import files to keep the index.js cleaner
