+++ my-app/src/book.js
@@ -1,0 +1,32 @@
+import React from 'react'
+
+const Book = ({ img, title, author, altText }) => {
+  // attribute, eventHandler
+  // onClick, onMouseOver
+  const clickHandler = () => {
+    alert('Hello World');
+  };
+  const complexExample = (author) => {
+    console.log(author);
+  };
+  return (
+    <article
+      className="book"
+      onMouseOut={() => {
+        console.log(title);
+      }}
+    >
+      <img src={img} alt={altText} />
+      <h1 onClick={() => console.log(title)}>{title}</h1>
+      <h4>{author}</h4>
+      <button type="button" onClick={clickHandler}>
+        Reference Example
+      </button>
+      <button type="button" onClick={() => complexExample(author)}>
+        more complex example
+      </button>
+    </article>
+  );
+};
+
+export default Book
