+++ my-app/src/books.js
@@ -1,0 +1,25 @@
+
+export const data = [
+  {
+    id: 1,
+    img: 'https://m.media-amazon.com/images/I/81TFRojp+uL._AC_UY218_.jpg',
+    title: 'Wheel of Time Box Set',
+    author: 'Robert Jordan / Brandon Sanderson',
+    altText: 'Wheel of Time Boxset',
+  },
+
+  {
+    id: 2,
+    img: 'https://m.media-amazon.com/images/I/51WEFBk27LL._AC_UY218_.jpg',
+    title: 'Options Trading For Dummies',
+    author: 'by Joe Duarte',
+    altText: 'Options Trading for Dummies',
+  },
+  {
+    id: 3,
+    img: 'https://m.media-amazon.com/images/I/81H03VyZLSL._AC_UY218_.jpg',
+    title: 'Technical Analysis For Dummies',
+    author: 'by Barbara Rockerfella',
+    altText: 'Technical Analysis For Dummies',
+  },
+];
